+++ R/bobyqa.R
@@ -87,7 +87,7 @@ bobyqaControl <- function(npt=NULL,
                           addProp = c("combined2", "combined1"),
                           calcTables=TRUE, compress=FALSE,
                           covMethod=c("r", ""),
-                          adjObf=TRUE, ci=0.95, sigdig=4, sigdigTable=NULL,
+                          adjObf=TRUE, ci=0.95, sigdig=3, sigdigTable=NULL,
                           eventSens=c("jump", "fd"), ...) {
 
   checkmate::assertIntegerish(npt, null.ok=TRUE, any.missing=FALSE, lower=2, len=1)