+++ R/newuoa.R
@@ -76,7 +76,7 @@ newuoaControl <- function(npt=NULL,
                           addProp = c("combined2", "combined1"),
                           calcTables=TRUE, compress=FALSE,
                           covMethod=c("r", ""),
-                          adjObf=TRUE, ci=0.95, sigdig=4, sigdigTable=NULL,
+                          adjObf=TRUE, ci=0.95, sigdig=3, sigdigTable=NULL,
                           boundedTransform=TRUE,
                           eventSens=c("jump", "fd"), ...) {